+++ examples/docs-examples/src/app/components/account-settings/page.tsx
@@ -1,0 +1,9 @@
+import { AccountSettings } from '@stackframe/stack';
+
+export default function MyAccountPage() {
+  return (
+    <AccountSettings
+      fullPage={true}
+    />
+  );
+}