+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,21 @@
 # e2e-tests
 
+## 2.6.11
+
+### Patch Changes
+
+- fixed account settings bugs
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.11
+
+## 2.6.10
+
+### Patch Changes
+
+- Various bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.10
+
 ## 2.6.9
 
 ### Patch Changes