+++ examples/demo/CHANGELOG.md
@@ -1,5 +1,25 @@
 # demo-app
 
+## 2.6.11
+
+### Patch Changes
+
+- fixed account settings bugs
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.11
+  - @stackframe/stack-ui@2.6.11
+  - @stackframe/stack@2.6.11
+
+## 2.6.10
+
+### Patch Changes
+
+- Various bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.10
+  - @stackframe/stack-ui@2.6.10
+  - @stackframe/stack@2.6.10
+
 ## 2.6.9
 
 ### Patch Changes