+++ examples/supabase/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @stackframe/example-supabase
 
+## 2.6.11
+
+### Patch Changes
+
+- fixed account settings bugs
+- Updated dependencies
+  - @stackframe/stack@2.6.11
+
+## 2.6.10
+
+### Patch Changes
+
+- Various bugfixes
+- Updated dependencies
+  - @stackframe/stack@2.6.10
+
 ## 2.6.9
 
 ### Patch Changes