+++ packages/stack-emails/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/stack-emails
 
+## 2.6.11
+
+### Patch Changes
+
+- fixed account settings bugs
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.11
+  - @stackframe/stack-ui@2.6.11
+
+## 2.6.10
+
+### Patch Changes
+
+- Various bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.10
+  - @stackframe/stack-ui@2.6.10
+
 ## 2.6.9
 
 ### Patch Changes